translation/requirements.txt
google-cloud-translate==3.0.2